requirements.txt
discord.py
youtube_dl
pynacl
pydub
pytube
python-dotenv
moviepy
youtube_search
requests